Shreyansh_Assessment_4/requirements.txt
aniso8601==10.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dotenv==0.9.9
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-RESTful==0.3.10
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
